1212/15/kjh/follow/feat get my followers/followings list API

+++ src/main/java/com/mycom/myapp/common/PagingResultDto.java
@@ -2,16 +2,15 @@ package com.mycom.myapp.common;
 
 import lombok.Getter;
 import lombok.NonNull;
-import lombok.Setter;
 
 import java.util.List;
 
 @Getter
 public class PagingResultDto<T> {
     private final List<T> content;
-    private final Integer totalCount;
+    private final Long totalCount;
 
-    public PagingResultDto(@NonNull List<T> content, @NonNull Integer totalCount) {
+    public PagingResultDto(@NonNull List<T> content, @NonNull Long totalCount) {
         this.content = content;
         this.totalCount = totalCount;
     }
